WIP: Really revoke callbacks to destructed objects, including locking and all.
If an object is destructed that requested events or callbacks from
libcurl, it now obtains the lock on the CurlEasyRequest object and then
revokes all callbacks; while the callbacks themselves also take the
lock on the CurlEasyRequest object. As a result it's no longer possible
to get callbacks to destructed objects, and such objects aren't
destructed while callbacks are being made to them.

All of this is really just an anti-crash hack of course; in reality LL
shouldn't even try to destroy these objects before we told libcurl we're
done with the handle and waited until libcurl phased them out...