Support texture delta for lossless/lossy + other changes.
Well, at least this showed me that changing the hash
algorithm isn't a problem. I added a database "version"
to allow us to enforce a database repair when ever a
hash changes again, which then automatically updates the
whole database with recalculated hash values.

I'm happy that I wrote the extra data (that I called 'Delta's)
on top of a BVH file that is needed to generate the asset
file (J2C) in an object oriented way, even though it was
the only one of its kind, because now I could rather easily
extend it to include textures as well. The TextureDelta
exists of only one boolean: lossless or not, but the principle
is the same as for animations (compare the Looping or not
boolean).

What did take me quite some time was to get the UI correct,
that wasn't as generic as it could have been. Most support
now is in LLFloaterNameDesc (the common base class of all
upload preview floaters).

Finally this large commit contains several other fixes and
improvements more or less related to this texture delta stuff.
More or less because they aren't really related but weren't
needed before and now they are :p. For example, support for
reading and writing booleans with AIXML. But also really
unrelated changes like changing <edit> into <singu> were I
added them in the first place and fixing of some typos,
documentation and float xml text.