Add BackEnd::is_know(LLUUID const&)
This function returns true if the id is a common UUID or already in the
database.

I had the following consideration while writing this patch:
Even though it's extremely unlikely that the locks (of database and/or
memory cache) are accessed by two viewers at the same time, or two
threads, I want to make it completely safe.
Therefore, to avoid deadlocks, the order of taking the database lock and
memory cache locks always has to be the same.
It makes sense to never try to obtain the database lock while already
holding the memory cache lock(s); the order is: first database lock and
then memory cache lock(s). This is because you normally want to obtain
the database lock at the start of a (long) transaction and only release
it at the end, while you want to keep the locks on the memory caches
only shortly.
Of course it is allowed to only obtain the memory cache lock, but if
then the database lock is needed, one first has to release the memory
cache lock(s). That means that the memory cache can change again,
theoretically (by another thread).
Therefore LockedBackEnd::getUploadedAsset(LLUUID const&) must start with
a re-check of the memory cache and keep its lock until after it added a
possible found upload to the memory cache.
The same held for the other getUploadedAsset functions.