Remove compiler warning about deprecated g_thread_init().
From the changelog of glib:

- libgthread is now an empty shell and g_thread_init() is no longer
   required (and has been deprecated)

As of 2.31, g_thread_supported() is a macro defined as (1), so
that g_thread_init() is never called anyway when compiled against
this version; however, despite that fact, the compiler still
produces a warning when parsing the code without this preprocessor
test.

This means that when the viewer is compiled against glibc 2.31 or
higher, then it won't run correctly when linked with a lower version
at runtime; and as long as we want to support older C libraries we
will have to compile against that older version for releases.

This patch makes no different with regard to that fact (without this
patch g_thread_init() would also never be called when compiled against
version 2.31 or higher because g_thread_supported() is a macro
returning 1, and not a function call). One could call that a bug in
glibc as it break unnecessarily backwards compatibility, but this is
out of our hands. It is not a problem however: we already (have to)
compile on the oldest distribution that we (want to) support.