Add debug setting CurlMaxPipelinedRequestsPerService
The maximum allowed total number of added easy handles per service
is much larger with pipeline support, so we need a second debug
setting next to CurlConcurrentConnectionsPerService for this.

This commit adds this new debug setting and propagates the per-grid
flag of whether pipelining is supported or not to
llmessage/aicurlperservice.cpp so that the maximum number of added
easy handles can be compared to either CurlConcurrentConnectionsPerService
or CurlMaxPipelinedRequestsPerService as a function of this pipeline
support.