Reapply "Sync LLAssetUploadResponders with upstream"
This is a manually applied e5e665b9eead26c0f47db3267a478d5ead93a5c1
the revert the revert. The main change is that we keep the
user callback of LLNewAgentInventoryResponder because singularity
needs to know if a snapshot was really uploaded successfully.
What LL did is just assume that "it will work" (or perhaps they
stall the whole viewer and don't do it in the background like we,
I didn't check).