Added HTTP pipelining detection heuristics.
The current heuristics turn on HTTP pipelining for
services that are used for textures and/or mesh when
they return a 'Keep-Alive:' header (to a GET or HEAD request).

Opensim currently does not return that header (and does
not support HTTP pipelining), so that works. However, if
in the future they'd start to support keeping alive connections
but NOT pipelining at the same moment - then we'd turn on
pipelining nevertheless; which is a bad thing (effectively
we'd drop to 1 connection per service and thus download textures
and mesh a lot slower). Hence, TODO: improve the heuristics.