HTTP pipelining: start with 1 connection for detection.
Instead of starting with blacklisting a site and then
removing that again when it turns out it can do pipelining,
start with http pipelining enabled but limit the number of
requests to 1. Then when it is detected that a site does
or does not do pipelining lift the connection limit and
optionally blacklist it.

This approach is necessary because libcurl 'remembers'
that a site was blacklisted EVEN when you remove that
again-- in the end resulting in two or more connections
to the service that have pipelining enabled, instead of 1.