Add MultiHandle::upload_finished_poll
After fixing libcurl, it closes connections a lot less,
reuses just a single connection for pipelining - and doesn't call the
socket callback as often anymore. As a result, the current 'upload
finished' detection stopped working: libcurl uses a quick poll(2) to
decide if it can write to a socket (and it always can) and then send GET
requests to the server internally. Absolutely NOTHING can be seen about
that from the outside, except that the pretransfer time on those
requests changed... Hence, we now regularly poll all added requests to
see if this variable was set and use that as our 'uploaded finished'
detection.