libcwd: specialized curl debug code.
This needs a specially hacked libcurl that abuses the socket callback
with an 'action' value of 1234567 to inform the viewer about its
internal connectdata structs, which subsequently are used to display
live graphical feedback in the HTTP console concerning possible
filedescriptor problems. File descriptors of active connections are
shown, followed by 'wr' when they need to be watched for writing or
reading respectively and '--' (or 'w-' or '-r') when not. While the
color of each ('w' and 'r') is red when the viewer isn't watching that
socket for writability and/or readablity.

Note that what followed is that libcurl *seldom* requires testing for
writability: it does that internally with a poll and just writes the
data without asking the application for it.