Bug fix in MergeIterator.
The result of this was that the same fd, when readable AND writable
was handled twice, which wasn't a problem. The only slightly negative
effect was that that caused 'ready', the return value of select(), to be
decremented by 2 extra, causing 1 or 2 other fds not to be handled that
loop (also not very serious).

I found this because the assert 'ready >= 0' fired, after substracting 2
from 'ready' while it had the value 1, which needs two fds to be active
at the same time: one being readable and writable and the other just
being readable or writable (happened maybe once per hour under heavy
load %-).