Better safe than sorry bug fix.
This might be pedantic, but I actually had an assert
fire (after a couple of months)... So, yeah - it can
happen, one in a million times or so? Not sure if SUCH
a small negative step can have a negative impact, but
better safe than sorry I guess. At least now I can remove
that assert from my own code.